app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
databricks-sql-connector==3.0.2